requirements.txt


fastapi
uvicorn
python-dateutil